+++ spec/fixtures/cookbooks/chef_client_updater_enterprise/resources/scheduler_reconvergence.rb
@@ -1,0 +1,1 @@
+../../../../../resources/scheduler_reconvergence.rb